pyproject.toml
[project]
name = "my-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.12.2",
    "meilisearch>=0.37.0",
    "python-dotenv>=1.1.1",
]
